requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
pydantic==2.9.2
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.34
alembic==1.13.2
httpx==0.27.2
python-dotenv==1.0.1
supabase==2.10.0
aiosmtplib==3.0.1
apscheduler==3.10.4
jinja2==3.1.4

